Modifications to support Arduino core libraries.
These changes are to implement compatibility between the newest
arduino-1.0, and the older arduino-002x family of IDE and libraries.

With the change to arduino-1.0, Arduino has taken the unfortunate and
unexpected path of breaking compatibility with nearly everything that
potentially uses the Arduino platform with 3rd party libraries and
shields.

The Wirefree library now supports arduino-0022, arduino-0023, and
arduino-1.0 seamlessly, without any user intervention.  Support for
IDEs < arduino-0022 is not guaranteed.